fix: move ecpayfunction to api file
